using intent change activity - team color and face
