+++ app/src/main/java/com/example/temicurling/MainActivity.java
@@ -1,5 +1,6 @@
 package com.example.temicurling;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.widget.TextView;
@@ -14,9 +15,6 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
 public class MainActivity extends AppCompatActivity {
-    private String currentRobotId = null;
-    private ValueEventListener currentListener = null;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -31,56 +29,13 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void selectRobot(String robotId) {
-        currentRobotId = robotId;
-        // Update the RobotController or Firebase listener to the new robotId
-        // You might need to reinitialize or update the existing RobotController
-        setupFirebaseListeners(robotId);
-    }
-    private void setupFirebaseListeners(String robotId) {
-        // Remove the old listener if it exists
-        if (currentListener != null && currentRobotId != null) {
-            DatabaseReference oldRef = FirebaseDatabase.getInstance().getReference("temi/" + currentRobotId);
-            oldRef.removeEventListener(currentListener);
-        }
+        Intent i = new Intent(getApplicationContext(), TemiFaceActivity.class);
+        i.putExtra("robotId", robotId);
+        startActivity(i);
 
-        // Reference to the new robot's data
-        DatabaseReference robotDataRef = FirebaseDatabase.getInstance().getReference("temi/" + robotId);
-
-        // Create a new ValueEventListener
-        currentListener = new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                String angleString = dataSnapshot.child("angle").getValue(String.class);
-                String distanceString = dataSnapshot.child("distance").getValue(String.class);
-                String outString = dataSnapshot.child("out").getValue(String.class);
-                String timeString = dataSnapshot.child("time").getValue(String.class);
-
-                if (angleString != null && distanceString != null && outString != null && timeString != null) {
-                    try {
-                        float angle = Float.parseFloat(angleString);
-                        float distance = Float.parseFloat(distanceString);
-                        boolean out = Boolean.parseBoolean(outString);
-                        long time = Long.parseLong(timeString);
-
-                        // Implement the logic to act on the data
-                        // ...
-
-                    } catch (NumberFormatException e) {
-                        Log.e("MainActivity", "Error parsing Firebase data for robot " + robotId, e);
-                    }
-                } else {
-                    Log.w("MainActivity", "Waiting for complete data for robot " + robotId);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-                Log.e("MainActivity", "Failed to read value for robot " + robotId + ": " + databaseError.toException());
-            }
-        };
-
-        // Attach the listener to the new reference
-        robotDataRef.addValueEventListener(currentListener);
-        currentRobotId = robotId;  // Update the current robot ID
+//        currentRobotId = robotId;
+//        // Update the RobotController or Firebase listener to the new robotId
+//        // You might need to reinitialize or update the existing RobotController
+//        setupFirebaseListeners(robotId);
     }
 }
